LayoutTests/fast/multicol/widows2.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
    description("Test column balancer behavior when widows requirements are high, and there are so many lines that there's no need for early breaks to honor widows");
</script>
<p>There should be 6 lines in the first column, 5 in the second and 5 in the last one.</p>
<div id="mc" style="-webkit-columns:3; columns:3; -webkit-column-rule:solid; line-height:50px;">
    <div style="orphans:1; widows:5;">
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
        line<br>
    </div>
</div>
<script>
    shouldBe("mc.offsetHeight", "300");
</script>
