third_party/blink/web_tests/platform/generic/fast/multicol/widows2-expected.txt

Test column balancer behavior when widows requirements are high, and there are so many lines that there's no need for early breaks to honor widows

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mc.offsetHeight is 300
PASS successfullyParsed is true

TEST COMPLETE


There should be 6 lines in the first column, 5 in the second and 5 in the last one.

line
line
line
line
line
line
line
line
line
line
line
line
line
line
line
line
